+++ app/tests/test_routes.py
@@ -1,0 +1,39 @@
+import os
+import unittest
+ 
+from app import app, db
+ 
+TEST_DB = 'test.db'
+ 
+class BasicTests(unittest.TestCase):
+ 
+  ############################
+  #### setup and teardown ####
+  ############################
+
+  # executed prior to each test
+  def setUp(self):
+    app.config['TESTING'] = True
+    app.config['WTF_CSRF_ENABLED'] = False
+    app.config['DEBUG'] = False
+    self.app = app.test_client()
+    db.drop_all()
+    db.create_all()
+
+    # Disable sending emails during unit testing
+    self.assertEqual(app.debug, False)
+
+  # executed after each test
+  def tearDown(self):
+    pass
+
+  ###############
+  #### tests ####
+  ###############
+
+  def test_main_page(self):
+    response = self.app.get('/', follow_redirects=True)
+    self.assertEqual(response.status_code, 200)
+
+if __name__ == "__main__":
+  unittest.main()
